Complete Assignment 6

+++ src/api/models/cat-model.js
@@ -16,12 +16,16 @@ const findCatById = async (id) => {
     "SELECT * FROM wsk_cats WHERE cat_id = ?",
     [id],
   );
+
   console.log("rows", rows);
+
   if (rows.length === 0) {
     return false;
   }
+
   return rows[0];
 };
+
 const findCatsByUserId = async (id) => {
   const [rows] = await promisePool.execute(
     `SELECT wsk_cats.*, wsk_users.name AS owner_name
@@ -34,41 +38,74 @@ const findCatsByUserId = async (id) => {
   console.log("rows", rows);
   return rows;
 };
+
 const addCat = async (cat) => {
   const { cat_name, weight, owner, filename, birthdate } = cat;
-  const sql = `INSERT INTO wsk_cats (cat_name, weight, owner, filename, birthdate)
+
+  const sql = `INSERT INTO wsk_cats
+               (cat_name, weight, owner, filename, birthdate)
                VALUES (?, ?, ?, ?, ?)`;
+
   const params = [cat_name, weight, owner, filename, birthdate];
+
   const rows = await promisePool.execute(sql, params);
+
   console.log("rows", rows);
+
   if (rows[0].affectedRows === 0) {
     return false;
   }
+
   return { cat_id: rows[0].insertId };
 };
 
-const modifyCat = async (cat, id) => {
-  const sql = promisePool.format(`UPDATE wsk_cats SET ? WHERE cat_id = ?`, [
-    cat,
-    id,
-  ]);
+const modifyCat = async (cat, id, user) => {
+  let sql;
+
+  if (user.role === "admin") {
+    sql = promisePool.format("UPDATE wsk_cats SET ? WHERE cat_id = ?", [
+      cat,
+      id,
+    ]);
+  } else {
+    sql = promisePool.format(
+      "UPDATE wsk_cats SET ? WHERE cat_id = ? AND owner = ?",
+      [cat, id, user.user_id],
+    );
+  }
+
   const rows = await promisePool.execute(sql);
+
   console.log("rows", rows);
+
   if (rows[0].affectedRows === 0) {
     return false;
   }
+
   return { message: "success" };
 };
 
-const removeCat = async (id) => {
-  const [rows] = await promisePool.execute(
-    "DELETE FROM wsk_cats WHERE cat_id = ?",
-    [id],
-  );
+const removeCat = async (id, user) => {
+  let rows;
+
+  if (user.role === "admin") {
+    [rows] = await promisePool.execute(
+      "DELETE FROM wsk_cats WHERE cat_id = ?",
+      [id],
+    );
+  } else {
+    [rows] = await promisePool.execute(
+      "DELETE FROM wsk_cats WHERE cat_id = ? AND owner = ?",
+      [id, user.user_id],
+    );
+  }
+
   console.log("rows", rows);
+
   if (rows.affectedRows === 0) {
     return false;
   }
+
   return { message: "success" };
 };
